Add logging individual episodes, remove log_seasons_by_numbers
This bit of abstraction is useful, but it does not match how the API works, and duplicating it for episodes would result in extra code. Furthermore, every call to (un)log_seasons_by_numbers previously resulted in re-fetching the show metadata, which is not ideal. This can be handled by the user instead.